Separate TestController to differnt abstraction
