+++ test/UnitTest/Microsoft.Test.AspNet.Common/AspNetCore/TestNonODataController.cs
@@ -1,0 +1,14 @@
+﻿// Copyright (c) Microsoft Corporation.  All rights reserved.
+// Licensed under the MIT License.  See License.txt in the project root for license information.
+
+using Microsoft.AspNetCore.Mvc;
+
+namespace Microsoft.Test.AspNet.OData
+{
+    /// <summary>
+    /// NonODataController is an abstracted ASP.NET [Core] controller.
+    /// </summary>
+    public class TestNonODataController : Controller
+    {
+    }
+}
