+++ test/UnitTest/Microsoft.Test.AspNet.Common/AspNetCore/TestODataController.cs
@@ -1,35 +1,14 @@
 ﻿// Copyright (c) Microsoft Corporation.  All rights reserved.
 // Licensed under the MIT License.  See License.txt in the project root for license information.
 
-#if NETCORE
 using System.Threading.Tasks;
 using Microsoft.AspNet.OData;
 using Microsoft.AspNet.OData.Results;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-#else
-using System.Net.Http;
-using System.Threading;
-using System.Threading.Tasks;
-using System.Web.Http;
-using System.Web.Http.Results;
-using Microsoft.AspNet.OData;
-using Microsoft.AspNet.OData.Results;
-#endif
 
 namespace Microsoft.Test.AspNet.OData
 {
-    /// <summary>
-    /// NonODataController is an abstracted ASP.NET [Core] controller.
-    /// </summary>
-#if NETCORE
-    public class TestNonODataController : Controller
-#else
-    public class TestNonODataController : ApiController
-#endif
-    {
-    }
-
     /// <summary>
     /// TestController is a controller designed to be used in UnitTests to abstract the controller
     /// semantics between AspNet and AspNet core. TestController implements (and hides) the convenience
@@ -52,37 +31,19 @@ namespace Microsoft.Test.AspNet.OData
         public new TestOkResult Ok() { return new TestOkResult(base.Ok()); }
 
         [NonAction]
-#if NETCORE
         public new TestOkObjectResult Ok(object value) { return new TestOkObjectResult(value); }
-#else
-        public new TestOkObjectResult<T> Ok<T>(T value) { return new TestOkObjectResult<T>(base.Ok<T>(value)); }
-#endif
 
         [NonAction]
-#if NETCORE
         public new TestCreatedObjectResult<TEntity> Created<TEntity>(TEntity value)
         {
             return new TestCreatedObjectResult<TEntity>(new CreatedODataResult<TEntity>(value));
         }
-#else
-        public new TestCreatedObjectResult<TEntity> Created<TEntity>(TEntity value)
-        {
-            return new TestCreatedObjectResult<TEntity>(base.Created<TEntity>(value));
-        }
-#endif
 
         [NonAction]
-#if NETCORE
         public new TestUpdatedObjectResult<TEntity> Updated<TEntity>(TEntity value)
         {
             return new TestUpdatedObjectResult<TEntity>(new UpdatedODataResult<TEntity>(value));
         }
-#else
-        public new TestUpdatedObjectResult<TEntity> Updated<TEntity>(TEntity value)
-        {
-            return new TestUpdatedObjectResult<TEntity>(base.Updated<TEntity>(value));
-        }
-#endif
     }
 
     /// <summary>
@@ -172,7 +133,6 @@ namespace Microsoft.Test.AspNet.OData
     /// <summary>
     /// Wrapper for OkObjectResult
     /// </summary>
-#if NETCORE
     public class TestOkObjectResult : TestObjectResult
     {
         public TestOkObjectResult(object innerResult)
@@ -181,17 +141,7 @@ namespace Microsoft.Test.AspNet.OData
             this.StatusCode = 200;
         }
     }
-#else
-    public class TestOkObjectResult<T> : TestActionResult
-    {
-        public TestOkObjectResult(OkNegotiatedContentResult<T> innerResult)
-            : base(innerResult)
-        {
-        }
-    }
-#endif
 
-#if NETCORE
     /// <summary>
     /// Platform-agnostic version of action result.
     /// </summary>
@@ -225,29 +175,4 @@ namespace Microsoft.Test.AspNet.OData
         {
         }
     }
-#else
-    /// <summary>
-    /// Platform-agnostic version of action result.
-    /// </summary>
-    public interface ITestActionResult : IHttpActionResult { }
-
-    /// <summary>
-    /// Wrapper for platform-agnostic version of action result.
-    /// </summary>
-    public class TestActionResult : ITestActionResult
-    {
-        private IHttpActionResult innerResult;
-
-        public TestActionResult(IHttpActionResult innerResult)
-        {
-            this.innerResult = innerResult;
-        }
-
-        public Task<HttpResponseMessage> ExecuteAsync(CancellationToken cancellationToken)
-        {
-            return innerResult.ExecuteAsync(cancellationToken);
-        }
-
-    }
-#endif
 }
